+++ tests/testthat/test-proxy.R
@@ -102,18 +102,22 @@ test_that("digits is working", {
 
 test_that("colSds and rowSds are working", {
     mt <- rsparsematrix(100, 100, 0.01)
+    dimnames(mt) <- list(paste0("row", seq_len(nrow(mt))),
+                         paste0("col", seq_len(ncol(mt))))
     expect_equal(rowSds(mt), apply(mt, 1, sd))
     expect_equal(colSds(mt), apply(mt, 2, sd))
-    expect_error(rowSds(matrix(mt)), "x must be a sparseMatrix")
-    expect_error(colSds(matrix(mt)), "x must be a sparseMatrix")
+    expect_equal(rowSds(as.matrix(mt)), apply(mt, 1, sd))
+    expect_equal(colSds(as.matrix(mt)), apply(mt, 2, sd))
 })
 
 test_that("colZeros and rowZeros are working", {
     mt <- rsparsematrix(100, 100, 0.01)
+    dimnames(mt) <- list(paste0("row", seq_len(nrow(mt))),
+                         paste0("col", seq_len(ncol(mt))))
     expect_equal(rowZeros(mt), apply(mt, 1, function(x) sum(x == 0)))
     expect_equal(colZeros(mt), apply(mt, 2, function(x) sum(x == 0)))
-    expect_error(rowZeros(matrix(mt)), "x must be a sparseMatrix")
-    expect_error(colZeros(matrix(mt)), "x must be a sparseMatrix")
+    expect_equal(rowZeros(as.matrix(mt)), apply(mt, 1, function(x) sum(x == 0)))
+    expect_equal(colZeros(as.matrix(mt)), apply(mt, 2, function(x) sum(x == 0)))
 })
 
 test_that("diag is working", {